refactor/EV-780/Refactoring file with componetization
