+++ lib/data/crypto_list_repository.dart
@@ -1,11 +1,12 @@
 import '../model/crypto_list_model.dart';
+import '../shared/providers/asset_provider.dart';
 
 class CryptoListRepository {
   List<CryptoListModel> cryptoListRepository = [
     CryptoListModel(
       shortName: 'BTC',
       fullName: 'Bitcoin',
-      cryptoLogo: 'assets/img/btc.png',
+      cryptoLogo: btcLogo,
       profitability: 75,
       userBalance: 7996,
       exchange: 114577.93,
@@ -13,7 +14,7 @@ class CryptoListRepository {
     CryptoListModel(
       shortName: 'ETH',
       fullName: 'Ethereum',
-      cryptoLogo: 'assets/img/eth.png',
+      cryptoLogo: ethLogo,
       profitability: 75,
       userBalance: 6557,
       exchange: 8960.35,
@@ -21,7 +22,7 @@ class CryptoListRepository {
     CryptoListModel(
       shortName: 'LTC',
       fullName: 'Litecoin',
-      cryptoLogo: 'assets/img/ltc.png',
+      cryptoLogo: ltcLogo,
       profitability: -0.7,
       userBalance: 245,
       exchange: 318.07,
