+++ lib/main.dart
@@ -3,7 +3,6 @@ import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-
 void main() {
   runApp(
     ProviderScope(
