+++ lib/shared/constants/app_colors.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+const Color kDefaultRed = Color(0xFFE02B57);
+const Color kDefaultGrey = Color(0xff757680);
+const Color kDefaultBlack = Color(0xff2F2F33);
