+++ lib/shared/providers/asset_provider.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/widgets.dart';
+
+const AssetImage warrenIcon = AssetImage('assets/img/warren.png');
+const AssetImage cryptoCurrencyIcon =
+    AssetImage('assets/img/cryptocurrency.png');
+const String btcLogo = 'assets/img/btc.png';
+const String ethLogo = 'assets/img/eth.png';
+const String ltcLogo = 'assets/img/ltc.png';
